requirements-alt.txt
apache-airflow-providers-amazon==3.2.0
apache-airflow-providers-apache-spark==2.1.3
apache-airflow-providers-celery==2.1.3
apache-airflow-providers-cncf-kubernetes==3.0.0
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-postgres==4.1.0
apache-airflow-providers-redis==2.0.4
apache-airflow-providers-snowflake==2.7.0
apache-airflow-providers-sqlite==2.1.3
attrdict==2.0.1
awscliv2==2.2.0
boto3==1.21.29
dbt-snowflake==1.1.0
humanfriendly==10.0
lxml==4.8.0
msal==1.17.0
paramiko==2.10.3
pendulum==2.1.2
pyarrow[pandas]==8.0.0
pylint-airflow
pytest==7.1.2
python-gnupg==0.4.9
s3fs==0.4.2
sqlfluff
